Add interactive Plotly visualizations, NTA temperature corrections, and UI improvements
Features added:
- Interactive Plotly plots for FCS and NTA analysis (zoom, pan, hover)
- NTA temperature/viscosity corrections (Stokes-Einstein equation)
- Cross-comparison visualization module
- Tab state persistence for cached analysis results
- Collapsible theoretical model preview
- Fixed non-unique index styling error in NTA
- Fixed anomaly_method NameError
- Added kaleido error handling for image export

New files:
- src/physics/nta_corrections.py (viscosity calculations)
- src/visualization/interactive_plots.py (Plotly functions)
- src/visualization/cross_comparison.py (FCS vs NTA comparison)
- docs/planning/GAP_ANALYSIS.md (progress tracking)

UI improvements:
- Better dashboard layout with proper spacing
- Always-visible modebar for plot controls
- Scroll zoom and hover tooltips enabled
- Reset Tab buttons for all analysis tabs

+++ src/visualization/anomaly_detection.py
@@ -232,13 +232,13 @@ class AnomalyDetector:
                 continue
             
             # Calculate Z-scores
-            mean = np.mean(valid_data)
-            std = np.std(valid_data)
+            mean = float(np.mean(valid_data))
+            std = float(np.std(valid_data))
             
             if std == 0:
                 continue
             
-            z_scores = np.abs((channel_data - mean) / std)
+            z_scores = np.abs((channel_data.astype(float) - mean) / std)
             outliers = z_scores > threshold
             
             data_copy.loc[outliers, 'is_outlier'] = True
